requirements.txt
flask==2.3.3
ultralytics==8.0.192
torch
opencv-python-headless==4.8.0.76
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0
seaborn==0.13.2
